library/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
